p38 implement add comment exception handling

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -1,8 +1,6 @@
 package life.majiang.community.controller;
 
 import life.majiang.community.dto.QuestionDTO;
-import life.majiang.community.mapper.QuestionMapper;
-import life.majiang.community.mapper.UserMapper;
 import life.majiang.community.model.Question;
 import life.majiang.community.model.User;
 import life.majiang.community.service.QuestionService;
@@ -14,7 +12,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 
 /**
@@ -30,7 +27,7 @@ public class PublishController {
 
   /** 修改页面 */
   @GetMapping("/publish/{id}")
-  public String edit(@PathVariable(name = "id") Integer id,
+  public String edit(@PathVariable(name = "id") Long id,
                      Model model) {
     QuestionDTO question = questionService.getById(id);
     model.addAttribute("title", question.getTitle());
@@ -61,7 +58,7 @@ public class PublishController {
           @RequestParam(value = "title", required = false) String title,
           @RequestParam(value ="description", required = false) String description,
           @RequestParam(value ="tag", required = false) String tag,
-          @RequestParam(value = "id", required = false) Integer id,
+          @RequestParam(value = "id", required = false) Long id,
           HttpServletRequest request,
           Model model) {
